move sagemaker models and remove tests
